create file for the context and wrap it to the index file
