+++ todo-app-context/src/index.js
@@ -3,11 +3,14 @@ import ReactDOM from 'react-dom/client';
 import './index.css';
 import App from './App';
 import reportWebVitals from './reportWebVitals';
+import { TodoProvider } from './TodoContext';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
   <StrictMode>
+    <TodoProvider>
     <App />
+    </TodoProvider>
   </StrictMode>
 );
